// ui/src/app/model/action.ts
export class Action {
    action: string;
    phone: string;
    line: string;
    from: string;
    to: string;
    conversationName: string;
    fromLine: string;
    toLine: string;
    callVia: string;
    fromPhone: string;
    toPhone: string;
    value: string;
    lineIndex: any;
    inTrace: string;
    outTrace: string;
    filter: string;
    linestate: string;
    callstate: string;
    calltype: string;
    url: string;
    continueonfailure: string;
    httpmethod: string;
    messagebody: string;
    command: string;
    resultIn: string;
    expression: string;
    maxWait: string;
    parkCode: string;
    server: string;
    userId: string;
    apiName: string;
    tagParam: string;
    configurationParameter: string;
    configurationValue: string;
    configuration360Parameter: string;
    parameters: any;
    hotelingPassword: string;
    syncType: string;
    prefix: string;
    tailing: string;
    intermediateConvName: string;
    mwistatus: string;
    ledstatus: string;
    hookType: string;
    duration: string;
    easing: string;
    xCoordinat: string;
    yCoordinat: string;
    touchType: string;
    replaceText: string;
    configType: string;
    eventType: string;
    callerNumber: string;
    callerName: string;
    operator: string;
    resetType: string;
    comment: boolean;
    dialingType: string;
    password: string;
    mediaType: string;
    selectedMediaParam: string;
    selectedmediaValue: string;
    selectedFlashType: string;
    selectedMode1: string;
    callKey: string;
    commentMessage: string;
    forceAction: any;
    user: string;
    resourceGroup: string;
}